fix(ci): split oversized test files under 1500-line decomposition limit
